requirements.txt
transformers==4.41.1
torch==2.3.0
accelerate
sentencepiece
protobuf
pyzmq
